add some headers and stubs
- sa_flags is 32bit as per POSIX manpages
- stubs for dladdr and Dl_info
- SIGIOT = SIGABRT
- signal.h should show ucontext
- unistd.h should show consts/l.h
- cosmocc -v

+++ libc/calls/struct/sigaction.h
@@ -12,7 +12,7 @@ struct sigaction {
     sighandler_t sa_handler;
     sigaction_f sa_sigaction;
   };
-  uint64_t sa_flags;
+  uint32_t sa_flags;
   void (*sa_restorer)(void);
   sigset_t sa_mask;
 };
